Fixed integer overflow issue and removed unneeded test data

+++ HMNGasApp/HMNGasApp/HMNGasApp/ViewModel/UsagePageViewModel.cs
@@ -66,7 +66,7 @@ namespace HMNGasApp.ViewModel
 
         private string FormatValueLabel(float vl)
         {
-            return "" + ((Int32) vl);
+            return "" + ((Int64) vl);
             
         }
